+++ src/main.js
@@ -1,7 +1,5 @@
 import TripPresenter from './presenter/trip-presenter.js';
 import FilterPresenter from './presenter/filter-presenter.js';
-import TripInfoPresenter from './presenter/trip-info-presenter.js';
-import SortPresenter from './presenter/sort-presenter.js';
 
 import DestinationModel from './model/destination-model.js';
 import OfferModel from './model/offers-model.js';
@@ -15,19 +13,8 @@ const destinationsModel = new DestinationModel(mockService);
 const pointsModel = new PointModel(mockService);
 const offersModel = new OfferModel(mockService);
 
-const tripInfoPresenterElement = new TripInfoPresenter({
-  tripInfoContainer: bodyElement,
-  destinationsModel,
-  pointsModel
-});
-
 const filterPresenterElement = new FilterPresenter({
   filterContainer: bodyElement,
-  pointsModel
-});
-
-const sortPresenterElement = new SortPresenter({
-  sortContainer: bodyElement,
   pointsModel
 });
 
@@ -38,7 +25,5 @@ const tripPresenterElement = new TripPresenter({
   pointsModel
 });
 
-tripInfoPresenterElement.init();
 filterPresenterElement.init();
-sortPresenterElement.init();
 tripPresenterElement.init();
