+++ src/mock/point.js
@@ -3,7 +3,7 @@ import {getRandomInt, getRandomBulValue, getDate, getRandomArrayElement} from '.
 export const generatePoint = (offerType, destinationId, offerIds) => ({
   id: crypto.randomUUID(),
   basePrice: getRandomInt(),
-  dateFrom: getDate(false),
+  dateFrom: getDate(true),
   dateTo: getDate(true),
   destination: destinationId,
   isFavorite: getRandomBulValue(),
