+++ src/view/empty-list-view.js
@@ -1,0 +1,9 @@
+import AbstractView from '../framework/view/abstract-view.js';
+
+const createTipEventListViewTemplate = () => '<p class="trip-events__msg">Click New Event to create your first point</p>';
+
+export default class EmptyListView extends AbstractView{
+  get template() {
+    return createTipEventListViewTemplate();
+  }
+}
