+++ src/view/tip-event-list-view.js
@@ -2,7 +2,7 @@ import AbstractView from '../framework/view/abstract-view.js';
 
 const createTipEventListViewTemplate = () => '<ul class="trip-events__list"></ul>';
 
-export default class TipEventList extends AbstractView{
+export default class TipEventListView extends AbstractView{
   get template() {
     return createTipEventListViewTemplate();
   }
